test(cli): split test_session.py into 7 focused files (D1 PR-3 step 2)
Splits the 4435-LOC monolithic ``tests/unit/cli/test_session.py`` into
seven focused files matching the post-D1 PR-3 session-CLI surface:

* ``test_login.py`` — core ``notebooklm login`` (Playwright + URL validation)
* ``test_use.py`` — ``notebooklm use`` (context-setting + auth-aware errors)
* ``test_status_clear.py`` — ``status`` and ``clear``
* ``test_auth_subcommands.py`` — ``auth check / logout / refresh / inspect``
  and the ``--browser-cookies`` login path
* ``test_session_edge_cases.py`` — legacy edge cases + Windows-permission
  regressions
* ``test_login_multi_account.py`` — ``login --account`` / ``--all-accounts``
  / stale-account-metadata cleanup
* ``test_login_chromium_fanout.py`` — Chromium multi-user-profile fan-out
  (``auth inspect``, ``--all-accounts``, ``--account``, explicit-profile
  selector, boundary conditions; #571)

Shared fixtures + helpers (``runner``, ``mock_auth``, ``mock_context_file``,
``_multiaccount_rookiepy_mock``, ``_make_chromium_profile``,
``_chromium_fanout_setup``, ``_install_chromium_fanout_patches``,
``_read_account``, ``_account_exists``) live in
``tests/unit/cli/_session_helpers.py`` and are auto-discovered via
``pytest_plugins = ["unit.cli._session_helpers"]`` in the top-level
``tests/conftest.py``. The plugin spec is the only valid surface for
``pytest_plugins`` per pytest's documented contract, and the path is
relative to the pytest rootdir (``tests/``) so the module resolves
without requiring ``tests`` to be importable as a Python package.

Other adjacent fixups:

* ``src/notebooklm/cli/session.py`` — annotate the ``from .services.login
  import (...)`` block with ``# noqa: F401`` since the symbols are
  imported as patch-surface re-exports rather than for in-module use.
* ``tests/unit/cli/conftest.py::_disable_chromium_profile_fanout`` —
  migrate from the ADR-007-forbidden ``monkeypatch.setattr("notebooklm.
  X.Y", ...)`` string-target form to ``patch(...)`` context-manager,
  which raises ``AttributeError`` on missing targets (catching
  relocation drift immediately).
* ``tests/integration/cli_vcr/test_login_browser_cookies.py`` — retarget
  the two ``_read_browser_cookies`` patches from session-side to
  services-side, since after D1 PR-3 step 1 deleted the proxy block the
  services-internal call no longer resolves through the session.py
  binding.
* ``tests/_lint/test_no_forbidden_monkeypatches.py`` — remove the stale
  ``tests/unit/cli/test_session.py`` allowlist entry (file no longer
  exists; replaced by the seven split files, none of which need the
  exception).

Verification:
  uv run ruff format . && uv run ruff check . && \
    uv run mypy src/notebooklm --ignore-missing-imports && \
    NOTEBOOKLM_HOME=/tmp/nbml-empty-d1cli uv run pytest -n auto tests/unit tests/integration
  → 5362 passed, 20 skipped, 0 failed.
